Puesta en marcha de cambios de sistema y actualización

+++ database/migrations/2016_07_06_205004_add_softdeleteordencompra.php
@@ -27,9 +27,11 @@ class AddSoftdeleteordencompra extends Migration
      */
     public function down()
     {
-        Schema::table('orden_de_compras', function (Blueprint $table) {
-            //
-            $table->dropColumn('deleted_at');
-        });
+        if (Schema::hasColumn('orden_de_compras', 'deleted_at')) {
+            Schema::table('orden_de_compras', function (Blueprint $table) {
+                //
+                $table->dropColumn('deleted_at');
+            });
+        }
     }
 }
